# web/openAPI3.yaml
openapi: 3.0.3
info:
  title: miflow Web API
  description: |
    miflow Web 服务 REST API，提供设备管理（DDD - device domain）与工作流管理（DDD - workflow domain）。
    设备相关接口需先通过 OAuth 登录（/login）获取 token 后访问。
  version: 1.0.0
  contact:
    name: miflow

servers:
  - url: http://localhost:8123
    description: 本地开发（默认端口 8123）

tags:
  - name: devices
    description: 设备管理（需 OAuth 登录）
  - name: workflows
    description: 工作流管理

paths:
  /api/rooms:
    get:
      tags: [devices]
      summary: 房间与设备
      description: 获取家庭、房间及设备列表（按房间分组），需 OAuth 登录。
      operationId: roomsList
      responses:
        '200':
          description: 家庭、房间及设备列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HomeWithRooms'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/rooms/{roomId}/devices:
    get:
      tags: [devices]
      summary: 按房间获取设备列表
      description: 获取指定房间的设备列表（含 status），需 OAuth 登录。可选 homeId 区分多家庭。
      operationId: roomsDevicesList
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
        - name: homeId
          in: query
          description: 家庭 ID（多家庭时用于区分）
          schema:
            type: string
      responses:
        '200':
          description: 设备列表（含 status）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomDevice'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/devices:
    get:
      tags: [devices]
      summary: 设备列表
      description: 获取接入的米家设备列表，需 OAuth 登录。
      operationId: devicesList
      parameters:
        - name: name
          in: query
          description: 按名称过滤
          schema:
            type: string
        - name: getVirtual
          in: query
          description: 是否包含虚拟设备
          schema:
            type: boolean
            default: false
        - name: getHuami
          in: query
          description: 华米设备过滤（0=不过滤）
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: 设备列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/devices/{id}:
    get:
      tags: [devices]
      summary: 设备详情
      description: 获取单个设备详情，需 OAuth 登录。
      operationId: deviceGet
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: 设备详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/devices/{id}/spec:
    get:
      tags: [devices]
      summary: 设备 MIoT 规格
      description: 获取设备的 MIoT SPEC（用于控制 UI），需 OAuth 登录。
      operationId: deviceSpec
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: MIoT SPEC（JSON 格式）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelSpec'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/devices/{id}/status:
    get:
      tags: [devices]
      summary: 设备状态
      description: 获取设备当前状态（开关、亮度、音量等），需 OAuth 登录。根据设备型号返回 supported 能力列表。
      operationId: deviceStatus
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: 设备状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [devices]
      summary: 设置设备状态
      description: 设置设备状态（开关、亮度、音量、静音等），需 OAuth 登录。
      operationId: deviceSetStatus
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceStatusInput'
      responses:
        '200':
          description: 设置成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/devices/{id}/control:
    patch:
      tags: [devices]
      summary: 设备控制（PATCH）
      description: 通过 action 控制设备（toggle/set_on/set_brightness/set_volume），需 OAuth 登录。
      operationId: deviceControlPatch
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceControlPatchRequest'
      responses:
        '200':
          description: 控制成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [devices]
      summary: 设备控制（MIoT 命令）
      description: 通过 MIoT 命令文本控制设备（等价于 m CLI），需 OAuth 登录。
      operationId: deviceControl
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceControlRequest'
      responses:
        '200':
          description: 控制成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/workflows:
    get:
      tags: [workflows]
      summary: 工作流列表
      description: 获取所有工作流。
      operationId: workflowsList
      responses:
        '200':
          description: 工作流列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workflow'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [workflows]
      summary: 创建工作流
      description: 创建新工作流。
      operationId: workflowCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowInput'
      responses:
        '200':
          description: 创建成功，返回工作流
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/workflows/{id}:
    get:
      tags: [workflows]
      summary: 工作流详情
      description: 获取单个工作流。
      operationId: workflowGet
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '200':
          description: 工作流详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [workflows]
      summary: 更新工作流
      description: 更新指定工作流。
      operationId: workflowUpdate
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowInput'
      responses:
        '200':
          description: 更新成功，返回工作流
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [workflows]
      summary: 删除工作流
      description: 删除指定工作流。
      operationId: workflowDelete
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '204':
          description: 删除成功，无响应体
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/workflows/{id}/run:
    post:
      tags: [workflows]
      summary: 执行工作流
      description: 异步执行指定工作流，需 OAuth 登录。
      operationId: workflowRun
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '202':
          description: 已接受，工作流已开始执行
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

components:
  parameters:
    DeviceId:
      name: id
      in: path
      required: true
      description: 设备 ID（did）
      schema:
        type: string
    WorkflowId:
      name: id
      in: path
      required: true
      description: 工作流 ID
      schema:
        type: string

  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 未登录，请先访问 /login 完成 OAuth 授权
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    HomeWithRooms:
      type: object
      description: 家庭及其房间
      properties:
        home_id:
          type: string
          description: 家庭 ID
        home_name:
          type: string
          description: 家庭名称
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/RoomWithDevices'

    RoomWithDevices:
      type: object
      description: 房间及其设备
      properties:
        room_id:
          type: string
          description: 房间 ID
        room_name:
          type: string
          description: 房间名称
        devices:
          type: array
          items:
            $ref: '#/components/schemas/Device'

    Device:
      type: object
      description: 米家接入设备（与 m list 输出一致）
      properties:
        did:
          type: string
          description: 设备 ID
        model:
          type: string
          description: 型号，用于查询 miot-spec.org SPEC
        name:
          type: string
          description: 设备名称
        token:
          type: string
          description: 设备 token

    DeviceControlRequest:
      type: object
      required: [command]
      properties:
        command:
          type: string
          description: MIoT 命令文本（如 "set_properties 2,1,1" 或自然语言指令）

    DeviceStatus:
      type: object
      description: 设备当前状态（根据型号返回不同字段）
      properties:
        on:
          type: boolean
          description: 开关状态（switch/light/plug）
        brightness:
          type: integer
          description: 亮度 0-100（light）
        volume:
          type: integer
          description: 音量 0-100（speaker）
        mute:
          type: boolean
          description: 静音状态（speaker）
        occupancy:
          description:  occupancy 传感器状态
        supported:
          type: array
          items:
            type: string
          description: 支持的能力列表（on/brightness/volume/mute/occupancy）

    RoomDevice:
      type: object
      description: 房间内设备（含 status）
      properties:
        id:
          type: string
        did:
          type: string
        name:
          type: string
        model:
          type: string
        type:
          type: string
          enum: [light, switch, plug, speaker, device]
        room_id:
          type: string
        room_name:
          type: string
        home_id:
          type: string
        status:
          type: object
          properties:
            online:
              type: boolean
            value:
              description: 主状态值
            on:
              type: boolean
            brightness:
              type: integer
            volume:
              type: integer
            supported:
              type: array
              items:
                type: string

    DeviceControlPatchRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [toggle, set_on, set_brightness, set_volume]
        value:
          description: action 对应的值（set_on 为 bool，set_brightness/set_volume 为 0-100）

    DeviceStatusInput:
      type: object
      description: 设置设备状态的请求体（可选字段）
      properties:
        on:
          type: boolean
        brightness:
          type: integer
          minimum: 0
          maximum: 100
        volume:
          type: integer
          minimum: 0
          maximum: 100
        mute:
          type: boolean

    ModelSpec:
      type: object
      description: MIoT 型号规格（来自 miot-spec.org）
      properties:
        type:
          type: string
        description:
          type: string
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceSpec'

    ServiceSpec:
      type: object
      properties:
        iid:
          type: integer
        description:
          type: string
        type:
          type: string
        properties:
          type: array
          items:
            $ref: '#/components/schemas/PropSpec'
        actions:
          type: array
          items:
            $ref: '#/components/schemas/ActionSpec'
        events:
          type: array
          items:
            $ref: '#/components/schemas/EventSpec'

    PropSpec:
      type: object
      properties:
        iid:
          type: integer
        description:
          type: string
        format:
          type: string
        access:
          type: array
          items:
            type: string

    ActionSpec:
      type: object
      properties:
        iid:
          type: integer
        description:
          type: string
        in:
          type: array
        out:
          type: array

    EventSpec:
      type: object
      properties:
        iid:
          type: integer
        description:
          type: string
        arguments:
          type: array

    Workflow:
      type: object
      description: 工作流
      properties:
        id:
          type: string
          description: 工作流 ID
        name:
          type: string
          description: 名称
        description:
          type: string
          description: 描述
        steps:
          type: array
          items:
            $ref: '#/components/schemas/Step'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    WorkflowInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: 名称（必填）
        description:
          type: string
        steps:
          type: array
          items:
            $ref: '#/components/schemas/Step'

    Step:
      type: object
      description: 工作流步骤
      properties:
        type:
          type: string
          enum: [tts, play_url, miio, delay]
          description: 步骤类型
        label:
          type: string
        device:
          type: string
          description: 设备 ID（tts/play_url 等）
        text:
          type: string
          description: TTS 文本
        url:
          type: string
          description: 播放 URL（play_url）
        miio_text:
          type: string
          description: MIoT 命令文本（miio）
        duration_ms:
          type: integer
          description: 延迟毫秒数（delay）

    StatusOk:
      type: object
      properties:
        status:
          type: string
          example: ok

    RunStatus:
      type: object
      properties:
        status:
          type: string
          example: started
        id:
          type: string
          description: 工作流 ID

    Error:
      type: object
      properties:
        error:
          type: string
          description: 错误信息